refactor: Event방식의 제거 및 validator적용
