+++ src/main/java/kitchenpos/order/domain/ChangeTableValidator.java
@@ -1,0 +1,6 @@
+package kitchenpos.order.domain;
+
+public interface ChangeTableValidator {
+
+    void validateOrderableTable(Long id);
+}
