+++ src/main/java/kitchenpos/tableGroup/application/TableGroupService.java
@@ -1,17 +1,18 @@
 package kitchenpos.tableGroup.application;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kitchenpos.event.CheckTableCanChangeEvent;
-import kitchenpos.tableGroup.application.request.TableGroupRequest;
-import kitchenpos.tableGroup.application.response.TableGroupResponse;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.tableGroup.application.request.TableGroupRequest;
+import kitchenpos.tableGroup.application.response.TableGroupResponse;
 import kitchenpos.tableGroup.domain.TableGroup;
 import kitchenpos.tableGroup.domain.TableGroupRepository;
 
@@ -19,15 +20,15 @@ import kitchenpos.tableGroup.domain.TableGroupRepository;
 @Transactional
 public class TableGroupService {
 
+    private final OrderRepository orderRepository;
     private final OrderTableRepository orderTableRepository;
     private final TableGroupRepository tableGroupRepository;
-    private final ApplicationEventPublisher applicationEventPublisher;
 
-    public TableGroupService(OrderTableRepository orderTableRepository, TableGroupRepository tableGroupRepository,
-                             ApplicationEventPublisher applicationEventPublisher) {
+    public TableGroupService(OrderRepository orderRepository, OrderTableRepository orderTableRepository,
+                             TableGroupRepository tableGroupRepository) {
+        this.orderRepository = orderRepository;
         this.orderTableRepository = orderTableRepository;
         this.tableGroupRepository = tableGroupRepository;
-        this.applicationEventPublisher = applicationEventPublisher;
     }
 
     public TableGroupResponse create(TableGroupRequest request) {
@@ -55,8 +56,15 @@ public class TableGroupService {
                 .map(OrderTable::getId)
                 .collect(Collectors.toList());
 
-        applicationEventPublisher.publishEvent(new CheckTableCanChangeEvent(orderTableIds));
+        validateIsPossibleToChange(orderTableIds);
         tableGroup.ungroup();
+    }
+
+    public void validateIsPossibleToChange(List<Long> orderTableIds) {
+        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
+                orderTableIds, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new IllegalArgumentException("조리중이나 식사중인 주문이 존재합니다.");
+        }
     }
 
     private TableGroup findTableGroup(Long tableGroupId) {
